requirements.txt
requests>=2.31.0
pandas>=2.1.0
numpy>=1.24.0
ta-lib>=0.4.28
python-dotenv>=1.0.0
pytest>=7.4.0
ta>=0.10.0
python-binance>=1.0.19
plotly>=5.18.0
streamlit>=1.31.1
websocket-client>=1.6.0 